Move Shadow Portal to `components/`
